+++ src/scandeval/benchmark_modules/hf.py
@@ -68,7 +68,9 @@ from ..types import ExtractLabelsFunction
 from ..utils import (
     block_terminal_output,
     create_model_cache_dir,
+    get_bos_token,
     get_class_by_name,
+    get_eos_token,
     internet_connection_available,
 )
 from .base import BenchmarkModule
@@ -862,15 +864,28 @@ def load_tokenizer(
         if add_prefix:
             loading_kwargs["add_prefix_space"] = True
 
-    while True:
+    num_retries = 5
+    for _ in range(num_retries):
         try:
-            return AutoTokenizer.from_pretrained(model_id, **loading_kwargs)
+            tokenizer = AutoTokenizer.from_pretrained(model_id, **loading_kwargs)
+            break
         except (JSONDecodeError, OSError, TypeError):
             raise InvalidModel(f"Could not load tokenizer for model {model_id!r}.")
         except (TimeoutError, RequestError):
             logger.info(f"Couldn't load tokenizer for {model_id!r}. Retrying.")
             sleep(5)
             continue
+    else:
+        raise InvalidModel(
+            f"Could not load tokenizer for model {model_id!r} after {num_retries} "
+            "attempts."
+        )
+
+    # Ensure that BOS, EOS and PAD tokens are set
+    tokenizer.bos_token, tokenizer.bos_token_id = get_bos_token(tokenizer=tokenizer)
+    tokenizer.eos_token, tokenizer.eos_token_id = get_eos_token(tokenizer=tokenizer)
+
+    return tokenizer
 
 
 def get_torch_dtype(