+++ src/scandeval/benchmark_modules/vllm.py
@@ -62,7 +62,9 @@ from ..types import ExtractLabelsFunction
 from ..utils import (
     clear_memory,
     create_model_cache_dir,
+    get_bos_token,
     get_end_of_chat_token_ids,
+    get_eos_token,
     log_once,
     should_prompts_be_stripped,
 )
@@ -1028,6 +1030,9 @@ def load_tokenizer(
             "attempts."
         )
 
+    # Ensure that BOS, EOS and PAD tokens are set
+    tokenizer.bos_token, tokenizer.bos_token_id = get_bos_token(tokenizer=tokenizer)
+    tokenizer.eos_token, tokenizer.eos_token_id = get_eos_token(tokenizer=tokenizer)
     if tokenizer.pad_token_id is None:
         tokenizer.pad_token = tokenizer.eos_token